SwiftDynamicsAssignment/requirements.txt
asgiref==3.4.1
backcall==0.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
debugpy==1.3.0
decorator==5.0.9
Django==3.2.5
djangorestframework==3.12.4
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
mysqlclient==2.0.3
oauthlib==3.1.1
packaging==21.0
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
